+++ MetalLink.Shared/Provinces/ProvinceDto.cs
@@ -2,7 +2,7 @@ namespace MetalLink.Shared.Provinces;
 
 public sealed class ProvinceDto
 {
-    public int ProvinceId { get; set; }
+    public long ProvinceId { get; set; }
     public string ProvinceName { get; set; } = string.Empty;
     public string? ProvinceCode { get; set; }
